src/main/webapp/app/entities/centros-salud/delete/centros-salud-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, fakeAsync, inject, tick } from '@angular/core/testing';
import { HttpResponse, provideHttpClient } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { CentrosSaludService } from '../service/centros-salud.service';

import { CentrosSaludDeleteDialogComponent } from './centros-salud-delete-dialog.component';

describe('CentrosSalud Management Delete Component', () => {
  let comp: CentrosSaludDeleteDialogComponent;
  let fixture: ComponentFixture<CentrosSaludDeleteDialogComponent>;
  let service: CentrosSaludService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [CentrosSaludDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(CentrosSaludDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(CentrosSaludDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(CentrosSaludService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
